% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuandlyR.R
\name{QuandlyR}
\alias{QuandlyR}
\title{QuandlyR}
\usage{
QuandlyR()
}
\value{
Return either a csv file per instrument or a data frame. The csv file is stored in the working directory and the data frame in the globalenv.
}
\description{
QuandlyR is an RStudio addin designed to simplify the process of downloading data from Fred
}
\author{
Jonathan Regenstein
}
\seealso{
\code{Quandl}

\code{shiny}
}

